feat(#1330): renamed governance to agreements
