+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/change-entry-method/page.tsx
@@ -1,20 +1,20 @@
 import {
+  CreateProposalChangeEntryMethodForm,
   SidePanel,
-  CreateProposalChangeVotingMethodForm,
 } from '@hypha-platform/epics';
 import { Locale } from '@hypha-platform/i18n';
 import { notFound } from 'next/navigation';
-import { getDhoPathGovernance } from '../../../../@tab/governance/constants';
 import { Plugin } from '../plugins';
-import { PATH_SELECT_SETTINGS_ACTION } from '@web/app/constants';
+import { getDhoPathAgreements } from '../../../../@tab/agreements/constants';
 import { findSpaceBySlug } from '@hypha-platform/core/server';
 import { db } from '@hypha-platform/storage-postgres';
+import { PATH_SELECT_SETTINGS_ACTION } from '@web/app/constants';
 
 type PageProps = {
   params: Promise<{ lang: Locale; id: string }>;
 };
 
-export default async function CreateChangeVotingMethodPage({
+export default async function CreateChangeEntryMethodPage({
   params,
 }: PageProps) {
   const { lang, id } = await params;
@@ -24,12 +24,11 @@ export default async function CreateChangeVotingMethodPage({
 
   if (!spaceFromDb) notFound();
   const { id: spaceId, web3SpaceId, slug: spaceSlug } = spaceFromDb;
-
-  const successfulUrl = getDhoPathGovernance(lang as Locale, id);
+  const successfulUrl = getDhoPathAgreements(lang as Locale, id);
 
   return (
     <SidePanel>
-      <CreateProposalChangeVotingMethodForm
+      <CreateProposalChangeEntryMethodForm
         spaceId={spaceId}
         web3SpaceId={web3SpaceId}
         successfulUrl={successfulUrl}
@@ -37,8 +36,8 @@ export default async function CreateChangeVotingMethodPage({
         plugin={
           <Plugin
             spaceSlug={spaceSlug}
-            name="change-voting-method"
             web3SpaceId={web3SpaceId}
+            name="change-entry-method"
           />
         }
       />
