+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/change-voting-method/page.tsx
@@ -1,45 +1,46 @@
-import { SidePanel } from '@hypha-platform/epics';
+import {
+  SidePanel,
+  CreateProposalChangeVotingMethodForm,
+} from '@hypha-platform/epics';
 import { Locale } from '@hypha-platform/i18n';
 import { notFound } from 'next/navigation';
-import { IssueNewTokenForm } from '@hypha-platform/epics';
+import { getDhoPathAgreements } from '../../../../@tab/agreements/constants';
 import { Plugin } from '../plugins';
 import { PATH_SELECT_SETTINGS_ACTION } from '@web/app/constants';
-import { getDhoPathGovernance } from '../../../../@tab/governance/constants';
 import { findSpaceBySlug } from '@hypha-platform/core/server';
 import { db } from '@hypha-platform/storage-postgres';
 
 type PageProps = {
   params: Promise<{ lang: Locale; id: string }>;
-  searchParams: Promise<{ back?: string; backMenu?: string }>;
 };
 
-export default async function IssueNewTokenPage({
+export default async function CreateChangeVotingMethodPage({
   params,
-  searchParams,
 }: PageProps) {
   const { lang, id } = await params;
-  const { back = '', backMenu = '' } = await searchParams;
 
   // TODO: implement authorization
   const spaceFromDb = await findSpaceBySlug({ slug: id }, { db });
 
   if (!spaceFromDb) notFound();
+  const { id: spaceId, web3SpaceId, slug: spaceSlug } = spaceFromDb;
 
-  const { id: spaceId, web3SpaceId } = spaceFromDb;
-
-  const successfulUrl = back ? back : getDhoPathGovernance(lang as Locale, id);
-  const backUrl = back
-    ? backMenu
-    : `${successfulUrl}${PATH_SELECT_SETTINGS_ACTION}`;
+  const successfulUrl = getDhoPathAgreements(lang as Locale, id);
 
   return (
     <SidePanel>
-      <IssueNewTokenForm
+      <CreateProposalChangeVotingMethodForm
         spaceId={spaceId}
         web3SpaceId={web3SpaceId}
         successfulUrl={successfulUrl}
-        backUrl={backUrl}
-        plugin={<Plugin name="issue-new-token" />}
+        backUrl={`${successfulUrl}${PATH_SELECT_SETTINGS_ACTION}`}
+        plugin={
+          <Plugin
+            spaceSlug={spaceSlug}
+            name="change-voting-method"
+            web3SpaceId={web3SpaceId}
+          />
+        }
       />
     </SidePanel>
   );
