+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/default.tsx
