+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/proposal/[documentSlug]/loading.tsx
