+++ apps/web/src/app/[lang]/dho/[id]/@tab/agreements/constants.ts
@@ -1,0 +1,5 @@
+import { Locale } from '@hypha-platform/i18n';
+
+export const getDhoPathAgreements = (lang: Locale, id: string) => {
+  return `/${lang}/dho/${id}/agreements`;
+};
