--- apps/web/src/app/[lang]/dho/[id]/@tab/governance/constants.ts
@@ -1,5 +1,0 @@
-import { Locale } from '@hypha-platform/i18n';
-
-export const getDhoPathGovernance = (lang: Locale, id: string) => {
-  return `/${lang}/dho/${id}/governance`;
-};
