+++ apps/web/src/app/[lang]/dho/[id]/_components/clean-path.spec.ts
@@ -3,13 +3,13 @@ import { cleanPath } from './clean-path';
 describe('cleanPath', () => {
   const testCases = [
     {
-      input: '/en/dho/my-org/governance/something/else',
-      expected: '/en/dho/my-org/governance',
-      description: 'should clean path after governance tab',
+      input: '/en/dho/my-org/agreements/something/else',
+      expected: '/en/dho/my-org/agreements',
+      description: 'should clean path after agreements tab',
     },
     {
-      input: '/en/dho/my-org/governance/select-settings-action',
-      expected: '/en/dho/my-org/governance',
+      input: '/en/dho/my-org/agreements/select-settings-action',
+      expected: '/en/dho/my-org/agreements',
       description: 'should remove settings action after tab',
     },
     {
