+++ apps/web/src/app/[lang]/dho/[id]/_components/get-active-tab-from-path.ts
@@ -3,6 +3,6 @@ export function getActiveTabFromPath(pathname: string) {
   // Match the pattern /[lang]/dho/[id]/{activeTab}/ to extract activeTab
   const match = pathname.match(/\/[^/]+\/dho\/[^/]+\/([^/]+)/);
 
-  // Return the matched tab name or default to 'governance'
-  return match?.[1] || 'governance';
+  // Return the matched tab name or default to 'agreements'
+  return match?.[1] || 'agreements';
 }
