+++ apps/web/src/app/[lang]/dho/[id]/layout.tsx
@@ -17,7 +17,7 @@ import Image from 'next/image';
 import { Carousel, CarouselContent, CarouselItem } from '@hypha-platform/ui';
 import Link from 'next/link';
 import { getAllSpaces, findSpaceBySlug } from '@hypha-platform/core/server';
-import { getDhoPathGovernance } from './@tab/governance/constants';
+import { getDhoPathAgreements } from './@tab/agreements/constants';
 import { ActionButtons } from './_components/action-buttons';
 import {
   DEFAULT_SPACE_AVATAR_IMAGE,
@@ -135,7 +135,7 @@ export default async function DhoLayout({
             web3SpaceId={spaceFromDb.web3SpaceId as number}
             isSandbox={spaceFromDb.flags.includes('sandbox')}
             isDemo={spaceFromDb.flags.includes('demo')}
-            configPath={`${getDhoPathGovernance(
+            configPath={`${getDhoPathAgreements(
               lang,
               daoSlug,
             )}/space-configuration`}
@@ -157,7 +157,7 @@ export default async function DhoLayout({
                   >
                     <Link
                       className="flex flex-col flex-1"
-                      href={getDhoPathGovernance(lang, space.slug as string)}
+                      href={getDhoPathAgreements(lang, space.slug as string)}
                     >
                       <SpaceCard
                         description={space.description as string}
@@ -169,7 +169,7 @@ export default async function DhoLayout({
                         isSandbox={space.flags?.includes('sandbox') ?? false}
                         isDemo={space.flags?.includes('demo') ?? false}
                         web3SpaceId={space.web3SpaceId as number}
-                        configPath={`${getDhoPathGovernance(
+                        configPath={`${getDhoPathAgreements(
                           lang,
                           space.slug,
                         )}/space-configuration`}
