+++ apps/web/src/app/[lang]/my-spaces/@aside/create/page.tsx
@@ -12,7 +12,7 @@ import {
 } from '@hypha-platform/core/client';
 import { useConfig } from 'wagmi';
 import { Button } from '@hypha-platform/ui';
-import { getDhoPathGovernance } from '@web/app/[lang]/dho/[id]/@tab/governance/constants';
+import { getDhoPathAgreements } from '@web/app/[lang]/dho/[id]/@tab/agreements/constants';
 
 export default function AsideCreateSpacePage() {
   const { lang } = useParams();
@@ -33,7 +33,7 @@ export default function AsideCreateSpacePage() {
 
   React.useEffect(() => {
     if (progress === 100 && spaceSlug) {
-      router.push(getDhoPathGovernance(lang as Locale, spaceSlug));
+      router.push(getDhoPathAgreements(lang as Locale, spaceSlug));
     }
   }, [progress, spaceSlug]);
